mde-troubleshooting-agent/requirements.txt
# pydantic-settings==2.8.1
# tabulate==0.9.0
# google-auth==2.36.0
# requests==2.32.3
# google-adk==1.0.0
# google-cloud-aiplatform[adk,agent-engines]==1.93.0
# llama-index==0.12.*
# db-dtypes==1.4.2
pydantic-settings==2.8.1
tabulate==0.9.0
google-auth==2.36.0
requests==2.32.3
google-adk==1.0.0
google-cloud-aiplatform[adk,agent-engines]==1.93.0
llama-index>=0.11.0  # Or just 'llama-index' for the absolute latest
db-dtypes            # Let pip decide the latest compatible version
fastapi
uvicorn
python-dotenv
uvicorn[standard]
fastapi-cors